ExpansionPort config was not saved in snapshot files

+++ Emulator/Media/Snapshot.h
@@ -22,7 +22,7 @@ class C64;
 struct Thumbnail {
     
     // Image size
-    isize width, height;
+    i32 width, height;
     
     // Raw texture data
     u32 screen[Texture::height * Texture::width];
